+++ src/pyday/__init__.py
@@ -1,5 +1,4 @@
 from .ChangLanguage import *
 from .DataReader.DataReader import *
-# from .Data2PDF.Data2PDF import *
-from .Data2PDF.Data2PDF2 import *
+from .Data2PDF.Data2PDF import *
 from .DataVis.DataVis import *